backend/pyproject.toml
[project]
name = "recipe-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.2",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.1",
]

[tool.uv]
dev-dependencies = []
